Add human-readable serialization for tagged types
Modify generate_tagged macro to serialize tagged types as a map for
human-readable formats, in the form

    {
        "type": "<type>",
        "value": "<value>"
    }

where <value> is the serialization of the inner type, and <type> is the
human-readable type name that is passed as a newly-added parameter to
the generate_tagged macro.

Note: we cannot stringify!() the inner type name for the "type" entry
because multiple types are defined in terms of the same inner type (e.g.
Svn and MinSvn are both tagged Uint's). Alternatively, we _could_ use
the integer tag rather than type names, but those are not easily
interpreted, requiring that anyone looking at the serialization refers
to the implementation/the spec to interpret the tag number. Finally, we
_could_ stringify!() the outer (i.e. the tagged) type name, which _are_
unique, however this is incoherent, as the "type" field annotates the
"value" field (and so <type> should describe <value>, not the enclosing
structure); additionally, they can be unwieldy, and inconsistent with
each other (.e.g some are suffixed with "Type", others are not, some are
prefixed with "Tagged" others are not).

The non-human-readable (i.e CBOR) serialization is untouched.

Signed-off-by: setrofim <setrofim@gmail.com>