Digests fix algorithm + human-readable serialization
Up to this point, AlgLabel was used for alg field inside Digest. This
can be either an arbitrary string, or one of CoseAlgorithm (IANA COSE
registry) IDs. [CoRIM spec][1] specifies that the Digest algorithm must
be interpreted according to the [IANA Named Information Hash Algorithm
Registry][2].

This replaced AlgLabel with HashAlogrithm that ensures that
- only valid names from the Hash Alg. Registry can be specified (rather
  than arbitrary strings).
- integer IDs are interpreted according to the Hash Alg. Registry (it
  contains a much smaller set of algorithms than the COSE registry, and
  for algorithms that appear in both, the numerical IDs differ).
- When serializing to CBOR, the integer ID is preferred where possible,
  as is recommended by the CoRIM spec. When serializing to
  human-readable format, the readable string name is used.

This also adds human-readable serialization for Digest. It is encoded
according to [RFC 6920][3], with the algorithm string name and the
base 64 encoding of the hash delimited by a ";".

[1]: https://www.ietf.org/archive/id/draft-ietf-rats-corim-07.html#name-digest
[2]: https://www.iana.org/assignments/named-information/named-information.xhtml
[3]: https://www.rfc-editor.org/rfc/rfc6920.html#section-3

Signed-off-by: setrofim <setrofim@gmail.com>