String representation for CoseAlgorithm + remove AlgLabel
- Add a string representation for CoseAlgorithm that uses the Name
  column in the [COSE Algorithms register][1].
- Remove AlgLabel and use CoseAlgorithm inside CoseKey directly. Rather
  than allowing an arbitrary string or arbitrary integer, this only
  allows valid algorithms.

[RFC 8152][2] specifies that the algorithm value must be taken from the
[COSE Algorithms register][1]. The [section on the register][3] inside
the RFC, further specifies that the Value (and, specifically, not the
Name) field is used to identify the algorithm.

The Value field allows both string and integer values, and so the
CoseKey alg field is correspondingly specified as "tstr / int". However,

- all currently registered values are integers
- no Private Use range is specified for string values

This means that, in practice, no valid string values for CoseAlgorithm
are possible at this time. Because of this, CBOR serialization simply
uses i64 in all cases.

The non-normative JSON serialization uses the human-readable Name
instead.

[1]: https://www.iana.org/assignments/cose/cose.xhtml#algorithms
[2]: https://www.rfc-editor.org/rfc/rfc8152#section-7.1
[3]: https://www.rfc-editor.org/rfc/rfc8152#section-16.4

Signed-off-by: setrofim <setrofim@gmail.com>