Add CoseKty enum for use in kty field of CoseKey
Replace the generic Label with the new CoseKty enum as the type of the
kty field inside CoseKey. Rather than allowing arbitrary strings or
integers in the field, this now only allows specific key types.

kty field is defined by [RFC 8152][1] to contain an identifier for the
family of they key, which, in turn, determines what other fields should
be expected to be present inside the  COSE Key structure.

Valid values for this field are defined in the [COSE Key Types][1] IANA
registry. Despite the field allowing for either string or integer
values, all values currently defined are integers.

The name "CoseKty" was chosen for the enum because the more descriptive
"CoseKeyType" is already taken.

[1]: https://www.rfc-editor.org/rfc/rfc8152#section-7
[2]: https://www.iana.org/assignments/cose/cose.xhtml#key-type

Signed-off-by: setrofim <setrofim@gmail.com>