Replace Labels inside CoseKey key_ops with CoseKeyOperation
Add CoseKeyOperation enum that represents valid key operations defined
by [RFC 8152][1]. Use this to replace the generic Label inside key_ops
field of CoseKey. This restricts the field to containing only valide key
operations, rather than arbitrary strings or integers.

CoseKeyOperation serializes to strings for JSON, and ints for CBOR. When
deserializing, it accepts either insts or strings for both formats.

The exception is "MAC create" and "MAC verify" ops, which do not have a
valid string representation in CBOR, and  so must be ints (JSON allow
non-normative string names, as JSON serialization itself is
non-normative).

[1]: https://www.rfc-editor.org/rfc/rfc8152#section-7.1

Signed-off-by: setrofim <setrofim@gmail.com>