Fix CoRIM tag encoding
The spec defines CoRIM tags field variants as

    tagged-concise-swid-tag = #6.505(bytes .cbor concise-swid-tag)
    tagged-concise-mid-tag = #6.506(bytes .cbor concise-mid-tag)
    tagged-concise-tl-tag = #6.508(bytes .cbor concise-tl-tag)

This means that, e.g., tagged-concise-swid-tag is CBOR tag 505 wrapping
a byte string, which contain the CBOR-encoded concise-swid-tag
structure. However, we encoded it as  COBR tag 505 wrapping the
concise-swid-tag structure (without first encoding it as a bytes
string).

This change updates CBOR serialization to match the spec. JSON
serialization is left as-is, because it is non-canonical, and it would
significantly negatively impact readability if we encoded the structures
as base64 strings.

Signed-off-by: setrofim <setrofim@gmail.com>