SignedCorim and ConciseRimTypeChoice rework
This is a major rework of how signed CoRIMs are handled. Structures that
are duplicates of COSE types are removed and coset library is used to
provide COSE support.

SignedCorim, rather than being a COSE_Sign1 object, now contains it,
alongside the deserialized CorimMap and header values (CorimMetaMap,
etc).

A builder is used to both construct and sign the SignedCorim values,
ensuring that the COSE_Sign1 is in sync with the deserialized fields.

As with coset library, actual signing and signature validation is
delegated to external code via hooks. CoseKeyOwner, CoseSigner, and
CoseVerifier traits are defined for this purpose. CoseKeyOwner provides
a way of getting a CoseKey from whatever format the actual key is in.
This is used to do validation of the key against the COSE headers, as
defined by the spec. CoseSigner and CoseVerifier define hooks, similar
to those defined by coset, for signature creation and verification
respectively.

Some types and ConciseRimTypeChoice variants have been renamed for
consistency.

deserialization for ConciseRimTypeChoice has been fixed (due to a know
ciborium issue, tagged types inside enums cannot be deserialized in the
usual way.).

Convenience to_/from_cbor() methods have been added to
ConciseRimTypeChoice, TaggedSignedCorim, and TaggedUnsignedCorim to
remove the need for the calling code to invoke ciborium directly.
Analogously, to_/from_json() have been added to TaggedUnsignedCorim, and
ConciseRimTypeChoice (SignedCorim is a COSE wrapper, and so does not
have/need a JSON representation).

Signed-off-by: setrofim <setrofim@gmail.com>