Add CorimMetaMapBuilder
Add a builder for CorimMetaMap that collects CorimSignerMap and
ValidityMap fields. Since a lot of the fields are optional (the only
thing that _needs_ to be set in the meta is the signer name), this makes
it easier to construct a valid CorimMetaMap, without needing to worry
about its internal structure.

Signed-off-by: setrofim <setrofim@gmail.com>