Consistent new() implementation for builders
Implement new() for all builders, aligned with default(), as per rust
API best practices.

Note in some cases an implementation was already provided via deriving
Constructor, however that generate a new() that took an argument for
every field. This does not really make sense for builders as they use
chaining, so this was replaced with an implementation that takes no
arguments.

Signed-off-by: setrofim <setrofim@gmail.com>