Unambiguous ExtensionValue JSON respresentation
When encoding ExtensionValue::Bytes into JSON, prefix the base64-encoded
string with "[base64]:" to indicate that the value is actually Bytes,
not Text. When decoding, look for this prefix as an indication that the
following value should be base64-decoded into Bytes.

Up to this point, ExtensionValue would try to deserialize all strings as
bytes, and would only yield Text if base64 decoding failed. This made it
impossible to correctly decode JSON representations of certain values.

NOTE: it is impossible to represent strings that start with "[base64]:",
as they would be interpreted as encoded bytes.

Signed-off-by: setrofim <setrofim@gmail.com>