Implement deep copy for CoRIM types
CoRIM types use Cow's (copy-on-write smart pointers) for contained
strings. A Cow can be either borrowed or owned. When a borrowed Cow is
cloned, the result is also a borrowed Cow that borrows from the same
underlying source. This means that cloning CoRIM structures does not
fully copy them, as they will still contain references to strings in the
original.

In order to allow some structures (such as environments) to exceed the
lifetimes of the CoRIMs that originated them, add to_fully_owned()
method that behaves much like clone() but ensures that any contained
Cow's are converted to own their contents.

Note: explicit lifetimes are used as an indicator that the lifetime of
the returned value is decoupled from that of the method's object.
However, as there no formal lifetime bounds that necessitate this in
the code base itself, clippy gets confused and mistakenly thinks that
the explicit lifetimes can be elided. This requires suppressing
clippy::needless_lifetime warnings for to_fully_owned() methods.

Signed-off-by: setrofim <setrofim@gmail.com>