fix: switch to definite length encoding for maps
Switch to using definite length encoding when serializing structs. While
not required by the CoRIM spec itself, CoSERV, which is defined in terms
of some of the CoRIM structures, does require it. (Also, the Go CoRIM
library also expects definite encoding).

Note: this fixes an actual issue with the meta map, as that goes into
the protected COSE header, and thus _must_ use definite length encoding.

Signed-off-by: setrofim <setrofim@gmail.com>